+++ Nexus/Engine/Cursor.cs
@@ -17,8 +17,8 @@ namespace Nexus.Engine {
 
 		public static int MouseX { get { return Cursor.mouseState.X; } }
 		public static int MouseY { get { return Cursor.mouseState.Y; } }
-		public static ushort MouseGridX { get { return (ushort) Snap.GridFloor((ushort)TilemapEnum.TileWidth, Systems.camera.posX + Cursor.mouseState.X); } }
-		public static ushort MouseGridY { get { return (ushort) Snap.GridFloor((ushort)TilemapEnum.TileHeight, Systems.camera.posY + Cursor.mouseState.Y); } }
+		public static byte MouseGridX { get { return (byte) Snap.GridFloor((byte)TilemapEnum.TileWidth, Systems.camera.posX + Cursor.mouseState.X); } }
+		public static byte MouseGridY { get { return (byte) Snap.GridFloor((byte)TilemapEnum.TileHeight, Systems.camera.posY + Cursor.mouseState.Y); } }
 
 		public static MouseDownState LeftMouseState;		// NOTE: Can use mouseState.LeftButton if you only need ON/OFF.
 		public static MouseDownState RightMouseState;		// NOTE: Can use mouseState.RightButton if you only need ON/OFF.
